requirements.txt
asgiref==3.3.4
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2.2
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.1
django-environ==0.4.5
django-nine==0.2.4
djangorestframework==3.12.4
elasticsearch==7.12.1
elasticsearch-dsl==7.3.0
idna==2.10
pip==21.0.1
pycparser==2.20
pyOpenSSL==20.0.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
setuptools==52.0.0.post20210125
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
